fix: nova paginação no registro
