+++ pages/nei/catalogo.vue
@@ -1,8 +1,9 @@
 <template>
 <ModalItemDetails v-if="itemsCache.length > 0" :item_index="itemIndex" :item_route="currentRoute" :item_details="currentItem" />
 <ModalActionConfirm>
-	<template v-slot:title> Confirmar aceitação </template>
-	<template v-slot:text> 
+    <template v-slot:title> Confirmar aceitação </template>
+    <!--
+    <template v-slot:text> 
         <div v-if="loadItems[0] != null" class="d-block">
             <div class="d-flex">
                 <div class="d-block mb-2 pe-2">
@@ -31,33 +32,33 @@
                 <input class="form-control" style="width: 225px !important;" v-model="itemQtd" type="number" pattern="[0,9]*" oninput="this.value = this.value.replace(/[^0-9]/g, '');">
             </div>
         </div>
+    </template>-->
+    <template v-slot:buttons> 
+        <button data-bs-dismiss="modal" class="btn btn-light-alert text-light mx-2"> Cancelar </button>
+        <button @click="sendRequest()" data-bs-dismiss="modal" class="btn btn-secondary mx-2"> Enviar </button>
     </template>
-	<template v-slot:buttons> 
-		<button data-bs-dismiss="modal" class="btn btn-light-alert text-light mx-2"> Cancelar </button>
-		<button @click="sendRequest()" data-bs-dismiss="modal" class="btn btn-secondary mx-2"> Enviar </button>
-	</template>
 </ModalActionConfirm>
 <div class="table-container d-block mt-2">
     <button class="d-none searching-btn" data-bs-toggle="modal" data-bs-target="#itemDetailing"></button>
     <div class="sub-catalog bg-light mb-4 ps-2 pe-2">
         <h6 class="sub-catalog-title ps-2 d-flex align-items-center opacity-75">
             <IconsInformation class="me-2"/>
-            Descrição da página 
+            Descrição da página
         </h6>
         <p class="sub-catalog-text opacity-75">Nesta página temos todos os itens disponíveis do almoxarifado(itens esgotados devem ser cadastrados novamente). 
             Ademais, o cadastro de novos itens e reposição da quantidade de algum item já existente é feito pelo botão 
         <span class="border-bottom border-dark-success pb-1">Adicionar <IconsPlus style="margin-bottom: 0px;"  width="18px" height="18px"/></span></p>
-    </div>   
+    </div>
     <div class="table-box row d-block">
-        <div class="table-actions d-flex justify-content-between aling-items-center" style="margin-bottom: -2px !important;">
-        <div class="d-flex me-1">
-            <ButtonsFilter style="margin-top: 3px;"/>
-        </div>
-          <span class="d-flex align-items-center bg-primary table-searchbar">
-                <input v-model="searchInput" class="searchbar bg-light form-control" placeholder="Pesquisar"/>          
+        <div class="table-actions d-flex justify-content-between aling-items-center" style="margin-bottom: -1px !important;">
+            <div class="d-flex">
+                <ButtonsFilter v-if="uploadReloader === 1" style="margin-top: 3px;" />
+            </div>
+            <span v-if="itemsLoad" class="position-sticky d-flex align-items-center bg-primary table-searchbar">
+                <input id="tableSearch" v-model="searchInput" class="searchbar bg-light form-control" placeholder="Pesquisar"/>          
                 <IconsSearchGlass class="bg-primary text-light search-glass"/>
-          </span>
-      </div>
+            </span>
+        </div>
         <TablesTable>
             <template v-slot:header>
                 <tr style="border: 1px #D9D9D9 solid;">
@@ -70,7 +71,7 @@
                 </tr>
             </template>
             <template v-slot:content>
-            <tr v-if="loadItems[0] != null" v-for="item in loadItems" :key="item.index" :data-index="item.index">
+            <tr v-if="itemsCache.length > 0" v-for="(item, index) in itemsCache[cacheIndex]" :key="index" :data-index="index">
                <th class="border" scope="row">
                     <span>{{ item.name }}</span>
                </th>
@@ -85,40 +86,47 @@
                    <span>{{ item.quantity }}</span>
                 </th>
                <th class="border">
-                   <span>CADASTRO 2024-05-11 09:20:02 Luís Freitas</span>
+                   <span>{{ item.lastRecord.operation }} {{  item.lastRecord.creationDate.slice(0, 16) }} {{ item.lastRecord.user.name }}</span>
                </th>
-               <th class="border" wdith="5%">
-                <TooltipsRectangular class="toolTip" style=" right: 10.3%; margin-top: -50px;" :toolTipState="toolTipState[0][item.index] ? toolTipState[0][item.index] : false" :toolTipText="'Detalhes'"/>
-                <TooltipsRectangular class="toolTip" style=" right: 8.3%; margin-top: -50px;" :toolTipState="toolTipState[1][item.index] ? toolTipState[1][item.index] : false" :toolTipText="'Solicitar'"/>
-                <button @mouseover="toolTipState[0][item.index] = true" @mouseout="toolTipState[0][item.index] = false" class="my-0 ms-2 details-btn position-sticky table-btn btn btn-primary" :class="{'d-none': store.isMobile}" @click="showDetails(item.index)" data-bs-toggle="modal" data-bs-target="#itemDetailing">
-                        <IconsSearchGlass width="16px" height="16px"/>
-                </button>
-                <button @mouseover="toolTipState[1][item.index] = true" @mouseout="toolTipState[1][item.index] = false" class="my-0 details-btn position-sticky table-btn btn btn-secondary" :class="{'d-none': store.isMobile}"  @click="showConfirm(item.index)" data-bs-toggle="modal" data-bs-target="#actionConfirm">
-                    <IconsSolicitation width="16px" height="16px"/>
-                </button>
+               <th class="border" width="5%">
+                   <TooltipsRectangular class="toolTip" style=" right: 7.4%; margin-top: -50px;" :toolTipState="toolTipState[0][index] ? toolTipState[0][index] : false" :toolTipText="'Detalhes'"/>
+                   <TooltipsRectangular class="toolTip" style=" right: 6%; margin-top: -50px;" :toolTipState="toolTipState[1][index] ? toolTipState[1][index] : false" :toolTipText="'Solicitar'"/>
+                   <button @mouseover="toolTipState[0][index] = true" @mouseout="toolTipState[0][index] = false" class="my-0 ms-2 details-btn position-sticky table-btn btn btn-primary" :class="{'d-none': store.isMobile}"  @click="showDetails(index)" data-bs-toggle="modal" data-bs-target="#itemDetailing">
+                       <IconsSearchGlass width="18px" height="19px"/>
+                   </button>
+                   <button @mouseover="toolTipState[1][index] = true" @mouseout="toolTipState[1][index] = false" class="my-0 details-btn position-sticky table-btn btn btn-secondary" :class="{'d-none': store.isMobile}"  @click="showConfirm(index)" data-bs-toggle="modal" data-bs-target="#actionConfirm">
+                        <IconsSolicitation width="16px" height="16px"/>
+                    </button>
                </th>
             </tr>
-            <div v-else-if="loadItems.length == 0" class="search-empty position-absolute mt-5">
-                <p class="text-dark-emphasis fs-5 opacity-50">Nenhum Resultado Encontrado.</p>
-            </div>
-            <div v-else class="warning-text d-flex aling-items-center justify-content-center">
-                <p class="text-dark-emphasis fs-5 opacity-50">Inventário vazio.</p>
+            <div v-else-if="itemsCache.length === 0 && !initialLoading" class="search-empty position-absolute mt-5">
+                <p class="text-dark-emphasis fs-5 opacity-50">Nenhum item Encontrado.</p>
             </div>
         </template>
-    </TablesTable>
+        </TablesTable>
     </div>
     <div class="d-flex justify-content-between me-2 mt-2">
-        <span class="ms-2 pages-info">Quantidade de itens da página: {{ loadItems.length }}</span> 
-        <nav v-if="itemsCache.length > 0" aria-label="Page navigation" class="pagination">
+        <span v-if="itemsCache.length > 0" class="ms-2 pages-info">Quantidade de itens da página: {{ itemsCache[cacheIndex].length }}</span>
+        
+        <nav v-if="itemsCache.length > 0 && finded.length === 0" aria-label="Page navigation" class="pagination">
             <ul class="pagination">
                 <li class="page-item">
                     <button class="page-link bg-primary text-light" :class="{'bg-dark-emphasis disabled': pagination == 0}" id="backPageBtn" @click="backPage"><span aria-hidden="true">&laquo;</span></button>
                 </li>
+                <li class="page-item" :key="0">
+                    <button class="page-link text-light" @click="page(0)" :class="{'bg-primary': !pagesFocus[0], 'bg-secondary': pagesFocus[0]}">{{ 1 }}</button>
+                </li>
+                <li v-show="pagination > 1" class="page-item">
+                    <button class="page-link bg-primary text-light">...</button>
+                </li>
                 <li class="page-item" v-for="i in totalPages >= 3 ? range(1+paginationRet, 3+paginationRet) : range(1,totalPages)" :key="i-1">
                     <button class="page-link text-light" @click="page(i-1)" :class="{'bg-primary': !pagesFocus[i-1], 'bg-secondary': pagesFocus[i-1]}">{{ i }}</button>
                 </li>
-                <li v-show="totalPages > 3 && paginationRet < totalPages-3" class="page-item">
+                <li v-show="totalPages > 3 && paginationRet < totalPages-4" class="page-item">
                     <button class="page-link bg-primary text-light">...</button>
+                </li>
+                <li class="page-item" :key="totalPages-1">
+                    <button class="page-link text-light" @click="page(totalPages-1)" :class="{'bg-primary': !pagesFocus[totalPages-1], 'bg-secondary': pagesFocus[totalPages-1]}">{{ totalPages }}</button>
                 </li>
                 <li class="page-item">
                     <button class="page-link bg-primary text-light" :class="{'bg-dark-emphasis disabled': pagination == totalPages-1 || searchInput !== ''}" id="fowardPageBtn" @click="fowardPage"><span aria-hidden="true">&raquo;</span></button>
@@ -140,14 +148,12 @@ import { postRequest } from '../../services/requests/requestsPOST.ts';
 definePageMeta({
     layout: 'client'
 })
-
-
-const userStore = useUser();
-const popUpStore = usePopupStore();
+/*SETANDO STORES*/
+const userStore = useUser()
 const store = useStorageStore();
 const searchStore = useSearch();
 /*VARIÁVEIS ÚTEIS PARA REQUISITAR OS ITENS E FILTRÁ-LOS*/ 
-const paginationRet = ref(0)
+const paginationRet = ref(1)
 function range(start, end) {
   return Array.from({ length: end - start + 1 }, (_, index) => start + index);
 }
@@ -158,93 +164,131 @@ let queryParams = ref({
     sort: 'id,desc', 
     isInverted: false
 });
+
 //Aqui faço a requisição em si, também possui parâmetros de filtros, sendo o padrão o de últimos atualizados(como está no banco de dados)
-const sortedResponse = async (sort, isInverted, pagination, paginationInverted) => {
+const itemsCache = ref([]);
+const searchCache = ref([])
+const cacheIndex = ref(0);
+const totalPages = ref(0);
+const isSearching = ref(false);
+let finded = [];
+const itemsReq = async (sort, isInverted, pagination, loadRequest, paginationInverted) => {
+    if(searchInput.value != ''){
+        finded = [];
+        store.isReloadItems = true;
+        if(searchCache.value.length >= totalPages.value){
+            for(let i = 0; i < searchCache.value.length; i++){
+                for(let j = 0; j < searchCache.value[i].length; j++){
+                    if(searchCache.value[i][j].name.includes(searchInput.value)){
+                        finded.push(searchCache.value[i][j]);
+                    }
+                    if(finded.length >= 20){ 
+                        itemsCache.value.push(finded);
+                        return 0
+                    };
+                }
+                if(finded.length >= 20){ 
+                        itemsCache.value.push(finded);
+                        return 0
+                };
+            }
+            if(finded.length === 0){
+                itemsCache.value = [];
+                return 0;
+            }
+            itemsCache.value.push(finded);
+            return 0;
+        }
+        for(let i = 0; i < totalPages.value; i++){
+            const res = await getItems(userStore, i, sort);
+            searchCache.value.push(res.content);
+            for(let j = 0; j < res.content.length; j++){
+                if(res.content[j].name.includes(searchInput.value)){
+                    finded.push(res.content[j]);
+                }
+                if(finded.length >= 20){ 
+                    itemsCache.value.push(finded);
+                    return 0
+                };
+            }
+            if(finded.length >= 20){ 
+                itemsCache.value.push(finded);
+                return 0
+            };
+        }
+        if(finded.length === 0){
+            itemsCache.value = [];
+            return 0;
+        }
+        itemsCache.value.push(finded);
+        return 0;
+    }
     if(isInverted){
         const res = await getItems(userStore, paginationInverted, sort)
-        return res
+        totalPages.value = res.totalPages
+        invertedPagination.value = totalPages-1;
+        itemsCache.value.push(res.content);
+        return res.totalPages
     } 
     const res = await getItems(userStore, pagination, sort)
-    return res
+    totalPages.value = res.totalPages
+    invertedPagination.value = totalPages-1;
+    loadRequest ? cacheIndex.value++ : 0;
+    itemsCache.value.push(res.content);
+    return res.totalPages
 }; 
-let response = await sortedResponse('', false, pagination.value, 0);
-let totalPages = response.totalPages
-invertedPagination.value = totalPages-1;
 
-
-let itemsCache = ref([])
-let indexCount = 0;
-for(let i = 0; i < totalPages; i++){
-    const res = await sortedResponse(queryParams.value.sort, false, pagination.value+i)
-    res.content.map((item) => {
-        item.index = indexCount;
-        indexCount++;
-        itemsCache.value.push(item)
-    });
-}
-store.items = itemsCache.value;
-async function reloadItems(sort, isInverted, invertedPagination){
-    let indexcount = 0;
-    if(isInverted){
-        for(let i = totalPages-1; i >= 0; i--){
-            const res = await sortedResponse(sort, true, 0, i)    
-            res.content.map((item) => {
-                item.index = indexcount;
-                itemsCache.value[indexcount] = item
-                indexcount++;
-            });
-        }
-        store.items = itemsCache.value;
-        return 1
-    }
-    for(let i = 0; i < totalPages; i++){
-        const res = await sortedResponse(sort, isInverted, i, invertedPagination)
-        res.content.map((item) => {
-            item.index = indexcount;
-            itemsCache.value[indexcount] = item
-            indexcount++;
-        });
-    }
-    store.items = itemsCache.value;
-    return 1;
-};
 
 const searchInput = ref("");
-const loadItems = computed(() => {
-    let items = [];
-    let page = 20*pagination.value
-    let aux = page;
-    let index = 0;
-    let find = 0;
-    if(searchInput.value != ''){
-        do{
-            if(store.items[index].name.includes(searchInput.value)){
-               items.push(store.items[index])
-               find++;
-            }
-            index++;
-        } while(index < store.items.length)
-        return items
+const initialLoading = ref(true);
+let reqsIndexCache = [0];
+let typingTimer; 
+const debounceTime = 1000; 
+const itemsLoad = computed(async() => {
+    if(initialLoading.value === true){
+        await itemsReq(queryParams.value.sort, false, 0, false, queryParams.value.isInverted);
+        return 0;
     }
-    do{
-        items.push(store.items[aux])
-        aux++;
-    }while(aux < store.items.length && aux < 20*(pagination.value+1));
-    return items
+    if(searchInput.value != ''){
+        clearTimeout(typingTimer);
+        typingTimer = setTimeout(() => {
+            itemsReq(queryParams.value.sort, false, 0, false, queryParams.value.isInverted)
+        }, debounceTime);
+        isSearching.value = true;
+    }
+    if(searchInput.value === ''){
+        if(isSearching.value === true){
+            clearTimeout(typingTimer);
+            typingTimer = setTimeout(() => {
+                store.isReloadItems = true;
+                isSearching.value = false;
+            }, debounceTime-500);
+            finded = [];
+        }
+    }
+    for(let i = 0; i < reqsIndexCache.length; i++){
+        if(pagination.value === reqsIndexCache[i]){
+            cacheIndex.value = i;        
+            return 0;
+        }
+    }
+    itemsReq(queryParams.value.sort, false, pagination.value, true, queryParams.value.isInverted);
+    reqsIndexCache.push(pagination.value)
 })
 
+
 provide('setItemsFilter', (filter, inverted) => {
-    queryParams.value.sort = filter
-    queryParams.value.isInverted = inverted
-    reloadItems(queryParams.value.sort, queryParams.value.isInverted, invertedPagination.value)
+    queryParams.value.sort = filter;
+    queryParams.value.isInverted = inverted;
+    store.isReloadItems = true;
+    //reloadItems(queryParams.value.sort, queryParams.value.isInverted, invertedPagination.value)
 });
 //Variáveis que o front vai pegar em si
 const itemIndex = ref(0);
-const currentItem = computed(() => store.items[itemIndex.value]);
+const currentItem = computed(() => itemsCache.value[cacheIndex.value][itemIndex.value]);
+
 const currentRoute = useRoute().fullPath.split('/')[2];
 
-
-const toolTip = ref([false, false, false, false])
 
 /*CÓDIGO PARA PAGINAÇÃO EM SI E RENDERIZAÇÃO DOS ITENS*/
 let pagesFocus = ref([true]);
@@ -253,8 +297,39 @@ for(let i = 0; i < totalPages; i++){
 };
 let count = 0;
 
+
+const uploadReloader = computed(() => {
+    if(store.isReloadItems === true){
+        pagination.value = 0;
+        cacheIndex.value = 0;
+        itemsCache.value = [];
+        reqsIndexCache = [0]
+
+        itemsReq(queryParams.value.sort, false, 0, false, queryParams.value.isInverted);
+
+        paginationRet.value = 1;
+        initialLoading.value = false
+        store.isReloadItems = false;
+
+        pagesFocus.value[count] = false;
+        count = 0;  
+        pagesFocus.value[0] = true;
+        return 1
+    } 
+    return 1;
+})
 const page = (async (index) => {
-    paginationRet.value = index+1 >= totalPages || index <= 0 ? paginationRet.value : index-1;
+    if (index <= 1 || index > totalPages.value - 3) {
+        if (index === 0) {
+            paginationRet.value = index + 1;
+        } else if (index === totalPages.value - 1) {
+            paginationRet.value = index - 3;
+        } else {
+            paginationRet.value = paginationRet.value;
+        }
+    } else {
+        paginationRet.value = index - 1;
+    }
     pagination.value = index;
     if(queryParams.value.isInverted){
         if(index < invertedPagination.value){
@@ -269,7 +344,7 @@ const page = (async (index) => {
     pagesFocus.value[count] = true;
 });
 const fowardPage = (async () => {
-    paginationRet.value = paginationRet.value < totalPages-3 ? paginationRet.value+1 : paginationRet.value  
+    paginationRet.value = pagination.value <= 1 || pagination.value >= totalPages.value-3 ? paginationRet.value : paginationRet.value+1  
     pagination.value++;
     if(queryParams.value.isInverted){
         invertedPagination.value--;
@@ -282,7 +357,7 @@ const fowardPage = (async () => {
     document.getElementById("backPageBtn").classList.remove("bg-dark-emphasis");
 });
 const backPage = (async () => {
-    paginationRet.value = paginationRet.value <= 0 ? paginationRet.value : paginationRet.value-1
+    paginationRet.value = pagination.value <= 2 || pagination.value > totalPages.value-3 ? paginationRet.value : paginationRet.value-1  
     pagination.value--;
     if(queryParams.value.isInverted){
         invertedPagination.value++;
@@ -298,11 +373,9 @@ const backPage = (async () => {
 const showDetails = (index) => {
     itemIndex.value = index;
 }
-
 const showConfirm = (index) => {
     itemIndex.value = index;
 }
-const itemQtd = ref(0)
 const description = ref('Text')
 const sendRequest = async () => {
     try{
@@ -312,10 +385,10 @@ const sendRequest = async () => {
     }
     return popUpStore.throwPopup('Solicitação enviada', '#0B3B69')
 }
-
 const toolTipState = ref([[], []]);
 /*HOOKS PARA RESPONSIVIDADE E MODO MOBILE*/
 onMounted(async () => {
+    initialLoading.value = false;
     if(searchStore.itemSearch.searching){
         showDetails(searchStore.itemSearch.itemId);
         const searching = document.getElementsByClassName('searching-btn'); 
@@ -368,7 +441,7 @@ onUpdated(async () => {
     display: block;
 }
 .table-container{
-    margin-bottom: 29px;
+    margin-bottom: 149px;
     width: 100%;
     display: block !important;
 }
